% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedobjective.R
\name{as.stuartFixedObjective}
\alias{as.stuartFixedObjective}
\title{Convert empirical to fixed objective.}
\usage{
as.stuartFixedObjective(x)
}
\arguments{
\item{x}{An object of class \code{stuartEmpiricalObjective}.}
}
\value{
Returns an object of class \code{stuartFixedObjective}
}
\description{
Convert an empirical objective to a fixed version to be used in item-selection. Sensible for extracting values from random selections and then using them in empirical but static objective functions.
}
\seealso{
\code{\link{empiricalobjective}}, \code{\link{fixedobjective}}
}
\author{
Martin Schultze
}
