% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{hmap}
\alias{hmap}
\title{Simple heatmap}
\usage{
hmap(mat, col = "RdYlBu", rotate = TRUE, reverse_color = TRUE,
  title = "", ...)
}
\arguments{
\item{mat}{A matrix}

\item{col}{RColorBrewer color palette to use}

\item{rotate}{Rotate the matrix 90º before calling `image`}

\item{reverse_color}{Reverse the color scale}

\item{title}{Plot title}
}
\description{
Plot a simple heatmap without clustering or dendorgram
}
