% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfolds.R
\name{kfolds}
\alias{kfolds}
\title{Data selection for k-folds crossvalidation.}
\usage{
kfolds(data, k = 2, seed = NULL)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{k}{The k-folding parameter. Defaults to 2, for an even split.}

\item{seed}{A random seed. See \code{\link{Random}} for more details.}
}
\value{
Returns a list containing two \code{data.frame}s, called calibrate and validate. The first corresponds to the calibration sample, the second to the validation sample.
}
\description{
Split a \code{data.frame} into two subsets for k-folds crossvalidation.
}
\seealso{
\code{\link{mmas}}, \code{\link{crossvalidate}}
}
\author{
Martin Schultze
}
\concept{
ACO subtests
}
