% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{set_fonts}
\alias{set_fonts}
\title{Define fonts}
\usage{
set_fonts(set = TRUE)
}
\arguments{
\item{set}{Set the default font family to Helvetica. If FALSE, just loads the custom font families}
}
\description{
Define some custom font families
}
\examples{
library(stuart)
define_fonts()

}
