% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{stuart_colors}
\alias{stuart_colors}
\title{Colour schemes}
\usage{
stuart_colors(type = "three1")
}
\arguments{
\item{type}{Name of colour scheme to choose colours from}
}
\value{
Vector of character strings of colours in hexadecimal
}
\description{
Vectors of different colours for use in figures
}
