% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidate.R
\name{crossvalidate}
\alias{crossvalidate}
\title{Cross-Validate a Measurement Model}
\usage{
crossvalidate(selection, old.data, new.data, invariance = "congeneric",
  fitness.func = fitness, filename = "stuart", file.remove = TRUE, ...)
}
\arguments{
\item{selection}{An object of class \code{stuartOutput}.}

\item{old.data}{A \code{data.frame} of the calibration sample.}

\item{new.data}{A \code{data.frame} of the validation sample.}

\item{invariance}{The invariance between the calibration and the validation sample. Can be one of 'congeneric', 'weak', 'strong', 'strict', or 'full', with the first being the default. Currently 'full' is only functional when using Mplus.}

\item{fitness.func}{A function that converts the results of model estimation into a pheromone. If none is provided the default function \code{fitness} is used. This can be examined with \code{body(fitness)}.}

\item{filename}{The stem of the filenames used to save inputs, outputs, and data files when using Mplus. Defaults to "stuart".}

\item{file.remove}{A logical indicating whether to remove the generated Mplus input and output files. Ignored if lavaan is used.}
}
\value{
Returns a \code{data.frame} with 2 observations of 7 variables. The first observation is the fitness of the final model for the calibration sample. The second observation is the fitness of the model for the validation sample.
\item{pheromone}{The value of the pheromone function.}
\item{chisq}{The Chi-Square of the model.}
\item{df}{Degrees of freedom of the model.}
\item{p}{p-Value of the Chi-Square test of model fit.}
\item{rmsea}{Root-Mean-Square-Error of Approximation.}
\item{srmr}{Standardized Root Mean Residual.}
\item{crel}{A measure of composite reliability.}
}
\description{
Cross-validate a measurement model obtained via either \code{\link{mmas}} or \code{\link{bruteforce}}.
}
\examples{
data(fairplayer)
calib <- fairplayer[1:80,]
valid <- fairplayer[81:135,]
fs <- list(SI1=paste0('SI',1:10,'t1'),
 EM1=paste0('EM',1:8,'t1'))

\donttest{
item.selection <- mmas(calib,fs,1,3,ants=1,colonies=1)
crossvalidate(item.selection,valid)
}
}
\author{
Martin Schultze
}
\seealso{
\code{\link{mmas}}, \code{\link{bruteforce}}
}
\concept{
ACO subtests
}

