% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{fairplayer}
\alias{fairplayer}
\title{fairplayer Intervention Data (2009)}
\format{A data frame with 135 observations on 55 variables. The variable names consist of two letters indicating the construct (SI: social intelligence, EM: empathy), a number indicating the item number on the scale, and a "t" followed by a number indicating the measurement occasion.}
\source{
Bull, H., Schultze, M., Scheithauer, H. (2009) School-based prevention of bullying and relational aggression: The fairplayer.manual. European Journal of Developmental Science, 3:313-317.
}
\usage{
fairplayer
}
\description{
Self-reported empathy (8 item scale) and social intelligence (10 item scale) at three different occasions.
}
\keyword{datasets}

